+++ packages/remirror__react-editors/CHANGELOG.md
@@ -382,6 +382,15 @@
   - @remirror/styles@3.0.0-beta.0
   - @remirror/pm@3.0.0-beta.0
 
+## 1.0.39
+
+> 2023-11-12
+
+### Patch Changes
+
+- Updated dependencies [2c6456fbe]
+  - remirror@2.0.37
+
 ## 1.0.38
 
 > 2023-08-03